Restore numbering for TOC, add TOC to keyCommands (#16226)
Fixes #16059
Closes #16106

Summary of the issue:

Due to markdown limitations, numbered headings and a numbered table of contents was removed from the user and dev guide when migrating from t2t to markdown.
However, these can be restored with CSS.

Description of user facing changes

Restores numbered headings to the table of contents and the headings within the user guide and developer guide.

Added TOC to key commands guide. Numbered heading are avoided as they don't match the user guide and may be misleading.