modernize cert ssl (#16255)
Summary of the issue:
Updating the NVDA certificate relies on a private, i.e. unstable method from the native python ssl library

Description of user facing changes
None

Description of development approach
Use the requests library to fetch the certificate, which provides a stable and standard method of creating an untrusted connection to update the cert.