Add support BrailleHID  protocol to BrailleEdge device (#16279)
Summary of the issue:
Add support BrailleHID protocol to BrailleEdge device

Description of user facing changes
User can use BrailleEdge device using BrailleHID protocol. NVDA keys remain unchanged.

Description of development approach
Five new keys have been added. On the front edge, starting from the left, there is the leftCursor key, in the center there is the home key, and on the far right there is the rightCursor key. On the top surface, there are control and alt keys on both sides of the spacebar. Additional gesture key mappings can be added to the gesture map based on user demand.