Remove profile from dirty list when deleting it (#16344)
Closes #16343

Summary of the issue:
When a dirty (modified) profile is deleted and the config is then saved, an error is raised in the function that saves the configuration. This may lead to other dirty profiles not being saved.

Description of user facing changes
A rare case where profiles are not saved is fixed.

Description of development approach
Remove the profile from dirty profiles list when it is deleted.