HP Secure Browser support (#16377): Detect brchrome and flag it as not supporting IAccessible2.
Summary of the issue:
Users have to explicitly configure in NVDA advanced options to "use UIA
for Edge and other chromium based browsers", for NVDA to work with HP Secure Browser.

Description of user facing changes:
Users won't have to explicitly configure UIA for all chromium-based browsers when
they want to use HP Secure Browser

Description of development approach:
Detect browser with a process name of "brchrome" which represents HP Secure Browser
and flag it as not supporting IAccecssible2. In this scenario, UIA will be
used by default for HP SB.