Added ability for braille to show what NVDA is speaking (#15956)
fixes #15898

Summary of the issue:
Both in NVDA core, as well as some add-ons, have features that work with speech but don't work with braille

Description of user facing changes
Braille only users will be able to read NVDA speech in braille, thus get access to a lot more features then they would otherwise

Description of development approach
A new braille mode has been added, called "speech emulation". In this mode, NVDA will send all text that has been spoken since the last time speech was canceled, to the braille display. If the speech was canceled do to the user scrolling the braille display back or fourth, the speech output mode will treat it as if the speech was not canceled, so the user can read the text without erasing it accidentally. If braille.handler.rawText is grater then 100000 characters, it will not be sent, because in my experience, NVDA will freeze when that much text is sent to the braille display at once. The braille settings panel has been rearranged so that all settings not aplickable to speech emulation is grouped together. A new combo box has been added to the braille settings panel, and a new script with the gesture NVDA+alt+t have been added, that allows the user to change braille mode.