Indicate category and option name in error message when validating settings dialog (#16352)
Summary of the issue:
The settings dialog contains always more error message boxes appearing when the validation of a parameter fails: modes available in cycle speech mode command or in cycle sound split mode command and regex validation in advanced settings have been added recently.

These messages do not formally mention the panel, nor the option which they are associated with. This may be difficult for the user to know which option is causing the error when the user has been modifying various settings in various setting panels (categories) before validating with "OK" or "Apply".

Description of user facing changes
The validation error messages now indicate the category and the name of the option which is causing the error so that the user can go to this option and modify it.

Description of development approach
Create a common dedicated error validation message with appropriate placeholders to be filled.