Properly support MSVC 14.40 toolset for getting the universal CRT files (#16480)
Microsoft is about to introduce Visual Studio 2022 version 17.10 in the near future. They decided to change the minor toolset version number, so we're going from MSVC 14.39 to 14.40.
See this blog post for details.
Note that they'll stick with version 143 in the crt version string, which is very weird.

Description of user facing changes
Building NVDA on visual studio 2022 version 17.10 (to be released) will work correctly.

Description of development approach
Rather than expecting all toolset directories to start with 14.3, we expect them to start with the major version part only (i.e. 14).