Support aria-errormessage (#16411)
Closes #8318

Summary of the issue:
Currently, NVDA does not read error messages referenced with aria-errormessage and aria-invalid (or platform native invalid).

Description of user facing changes
Error messages referenced by aria-errormessage are reported in speech and Braille on nodes marked as invalid.

Description of development approach
NVDAObjects have a new property, errorMessage, which is retrieved via the error IA2 relation. For virtual buffers, this is done in the virtual buffer backend and stored in the virtual buffer node. For focus mode, this is done at call-time.