Correctly announce selection changes in inline cel edit control in Microsoft Excel (#16473)
Fixes #15843

Summary of the issue:
When editing a cell in Microsoft Excel with inline cell editing enabled, NvDA does not announce text selection changes.
This is because in order to support inline cell editing, NvDA needs to redirect focus to the real cell edit control - Excel fires MSAA focus on a broken control. However, caret events do not know that focus was redirected and only go to the last queued focus object, which is still the broken control.
Note that this already works when NVDA uses UIA to access MS Excel (as the UIA control is not broken).

Description of user facing changes
NVDA announces selection changes in the inline cel edit control in Microsoft Excel.

Description of development approach
Rather than redirecting focus events to an obj's focusRedirect property in executeEvent, do it earlier in queueEvent, so that lastQueuedFocusObject is the possibbly redirected focus and therefore other code paths in NVDA check against what the focus actually ends up being.