Fix reading the last empty line of a text in apps using Java Access Bridge (#16568)
Fixes #9376.

Summary of the issue:
In applications using Java Access Bridge, if the last line of a text is an empty line, then NVDA reads the previous line as the last line.

Description of user facing changes
NVDA will now correctly read the last empty line as "blank".

Description of development approach
The issue is related to fixes of #1892 in 7ef99c3, which work around the line bounds returned by JAB for the last character of a text. Specifically, the current behavior is to get the line bounds of the previous character, which works well for single-line text inputs. But if the text is multi-line, and we're on the last empty line, then getting the line bounds of the previous character results in repeating the penultimate line.

The fix is to separately handle the case of empty last line, which is determined by having "\n" as the previous character.