Update and migrate translating page from wiki (#16577)
NV Access is intending to move most of the wiki into the repository, so changes are tracked better and easier to propose.
The documentation for translating NVDA is very dated, particularly since the update of the translation system.

Description of user facing changes
In order to start the process of updating and migrating the translating documentation, the first page was moved from the wiki:
https://github.com/nvaccess/nvda/wiki/Translating

The following changes were made to update the contents:

the translation status links to Crowdin, which neatly tracks the status for the NVDA interface. Instructions for generating structured diffs to check the status of user docs will be linked when they are made.
warnings are added to note that docs are out of date in general, until all referenced pages are updated
links to the manual process are removed. The manual process was only ever documented for the interface, which has moved crowdin. References to this page will be removed.
references to text2tags were updated to markdown, such as removing references to the no longer needed locale.t2tconf
Additionally, some grammar and formatting fixes were made such as new lines
Removed "Missing information" now that this in the repo, normal issues/PRs apply