Disable style navigation quickNav in MS Word (#16595)
Fixes #16546

Summary of the issue:
Style navigation is too slow in MSWord even with UIA enabled.

Description of user facing changes
Message "not supported in this document" is spoken when trying style navigation in MSWord.

Description of development approach
I did a bit of refactoring. Instead of trying to catch all the conditions in browseMode.BrowseModeDocumentTreeInterceptor, I overrode _iterTextStyle() method in corresponding child subclasses.