Update contributing.md (#16621)
#16601

Summary of the issue:
Compiled a list of steps for developers to set up a development environment on their machines into the contributing.md document. Steps were adapted mainly from documents in the "Overview of contribution process" section.

Description of user facing changes
No changes for end-users. Because of this, I did not make any updates to the changes.md file, but if that was incorrect, please let me know and I will update it right away.

Description of development approach
Written using Github's built in editor to check the formatting and changes in real time. There were only changes to the documentation, no coding changes.