Remove generic anchor names from the User Guide (#16661)
Discussion in #16655

Summary of the issue:
Some paragraphs in the User Guide have inexplicit anchors names such as #toc45. In case they need to be referenced in the future, in the User Guide, in NVDA's GUI (context help) or in an external source, it would be desirable to have a more explicit name. Moreover, the auto-generated anchor names (#tocNNN) have probably changed various times in the past due to the evolution of the structure of the user guide.

Until NVDA 2024.1 (included), the base document for the User Guide was the t2t file and these anchors were "floating", depending on the evolution of the document. E.g. the paragraph "Open General settings" has anchor #toc109 in NVDA 2024.1 but #toc110 in NVDA 2024.2beta3.

Description of user facing changes
Replace auto-generated anchors #tocNNN by explicit anchor names.

Description of development approach
Commit 8891a1e
Renamed the anchors of the "open settings" paragraphs.
Note: These paragraphs were added in NVDA 2024.1 but their anchor names differ between 2024.1 and 2024.2beta3 due to changes in the User Guide structure.

Commit 9d726d6
Renamed the 4 other anchors that have changed between NVDA 2024.1 and NVDA 2024.2beta3.
Note: These anchor names have changed due to changes in the User Guide structure.

Commit ce8591f
Rename the two remaining anchors whose name has not changed between 2024.1 and 2024.2beta3.
Note: These anchors were first present in NVDA 2024.1; before, they were not present since they correspond to level 5 heading (or level 4) which were not listed in the table of contents of the t2t file.